roachtest: add acceptance/build-info
Move the build-info acceptance test to a new acceptance/build-info
roachtest. Move the existing cli/node-status to
acceptance/cli/node-status. The acceptance roachtests are all going to
share the same cluster to amortize the cluster setup/teardown time.

Fun fact: the acceptance/build-info roachtest is about twice as fast as
the old build-info acceptance test.

See #29151

Release note: None
