+++ pkg/cmd/roachtest/build_info.go
@@ -1,0 +1,50 @@
+// Copyright 2018 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied. See the License for the specific language governing
+// permissions and limitations under the License. See the AUTHORS file
+// for names of contributors.
+
+package main
+
+import (
+	"context"
+	"net/http"
+
+	"github.com/cockroachdb/cockroach/pkg/server/serverpb"
+	"github.com/cockroachdb/cockroach/pkg/util/httputil"
+)
+
+func runBuildInfo(ctx context.Context, t *test, c *cluster) {
+	c.Put(ctx, cockroach, "./cockroach", c.All())
+	c.Wipe(ctx)
+	c.Start(ctx)
+
+	var details serverpb.DetailsResponse
+	url := `http://` + c.ExternalAdminUIAddr(ctx, c.Node(1))[0] + `/_status/details/local`
+	err := httputil.GetJSON(http.Client{}, url, &details)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	bi := details.BuildInfo
+	testData := map[string]string{
+		"go_version": bi.GoVersion,
+		"tag":        bi.Tag,
+		"time":       bi.Time,
+		"revision":   bi.Revision,
+	}
+	for key, val := range testData {
+		if val == "" {
+			t.Fatalf("build info not set for \"%s\"", key)
+		}
+	}
+}
